Split tf-hparams-parallel-coords-plot.ts into multiple files.
