+++ tensorboard/plugins/hparams/tf_hparams_parallel_coords_plot/tf-hparams-parallel-coords-plot.html
@@ -13,12 +13,55 @@ distributed under the License is distributed on an "AS IS" BASIS,
 WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 See the License for the specific language governing permissions and
 limitations under the License.
-  -->
+-->
 
-<!-- Renders the parallel coordinates plot. Most of the heavy lifting is done
-     by tf-hparams-parallel-coords-plot.ts. See the documentation there for
-     more details.
-  -->
+<!--
+ * A D3-based implementation of a parallel-coordinates plot of the sessions
+ * groups.
+ *
+ * This tf-hparams-parallel-coords-plot element displays a collection of
+ * parallel vertical axes uniformally spaced with each axis corresponding to a
+ * column--either an hyperparameter or a metric. Additionally each session
+ * group is represented by a polyline whose vertices are points on the axes;
+ * the vertex on the axis corresponding to a given column is the point on that
+ * axis representing the column-value of the session group.
+ *
+ * Since the relations shown by the plot depends on the ordering of
+ * the axes, the element allows re-ordering of the axes by dragging the axes
+ * titles. For columns with a numeric domain (e.g. metrics), it allows
+ * setting the scale of an axis to be 'linear', 'logarithmic' or 'quantile'.
+ * The element also colors each line based on a given 'color-by' column value
+ * of the represented session group. The 'color-by' column can be selected
+ * to be any column with numeric domain.
+ *
+ * Finally, the user can filter the displayed session groups by "brushing"
+ * axes. By clicking and dragging the mouse pointer vertically along an axis
+ * the user can select part of the axes represented by a semi-transparent
+ * overlay on the axis. We call that rectangle the brush-selection for the
+ * axis. The user can cancel the brush-selection by clicking on the axis
+ * outside the selection. Brush selections "filter" the polylines that are
+ * displayed as follows. At any given time--for the axes that have a brush
+ * selection--only the polylines that have each of these axes' vertices be
+ * inside the axis' brush-selection are displayed colored; the rest
+ * of the polylines are grayed-out.
+ *
+ * The implementation is based on the following cooperating classes:
+ * + Defined in axes.ts:
+ *   + Axis. Represents a single axis. Defined in axes.ts
+ *   + AxesCollection. Represents the collection of axes. Responsible for
+ *     handling axis drag and re-ordering behavior.
+ * + Defined in lines.ts:
+ *   + LinesCollection. Manages the collection of lines representing the session
+ *     groups.
+ * + Defind in interaction_manager.ts:
+ *   + InteractionManager. Manages the interaction of entire plot with the user.
+ *     Contains event handlers that respond to events in the DOM (such as an
+ *     Axis being dragged) and calls appropriate methods in the other classes
+ *     to update their state and redraw the necessary parts of the plot in
+ *     response.
+ *
+ * See the individual class comments in the respective files for more details.
+-->
 <link rel="import" href="../polymer/polymer.html">
 <link rel="import" href="../tf-imports/d3.html">
 <link rel="import" href="../tf-imports/lodash.html">
@@ -89,7 +132,9 @@ limitations under the License.
     </style>
   </template>
 
-  <script src="tf-hparams-parallel-coords-plot.js"></script>
+  <script src="axes.js"></script>
+  <script src="lines.js"></script>
+  <script src="interaction_manager.js"></script>
   <script>
     Polymer({
       is: 'tf-hparams-parallel-coords-plot',
